Use uint256 instead of uint

+++ contracts/Nodes.sol
@@ -145,7 +145,7 @@ contract Nodes is Permissions, INodes {
         if (space > 0) {
             _moveNodeToNewSpaceMap(
                 nodeIndex,
-                (uint(spaceOfNodes[nodeIndex].freeSpace) - space).toUint8()
+                (uint256(spaceOfNodes[nodeIndex].freeSpace) - space).toUint8()
             );
         }
         return true;
@@ -165,7 +165,7 @@ contract Nodes is Permissions, INodes {
         if (space > 0) {
             _moveNodeToNewSpaceMap(
                 nodeIndex,
-                (uint(spaceOfNodes[nodeIndex].freeSpace) + space).toUint8()
+                (uint256(spaceOfNodes[nodeIndex].freeSpace) + space).toUint8()
             );
         }
     }
